\name{combine.character}
\alias{combine.character}
\title{\code{\link{combine}} character}
\description{\code{\link{combine}} character strings by pasting them together}
\usage{combine.character(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





