\name{combine.NULL}
\alias{combine.NULL}
\title{\code{\link{combine}} NULL}
\description{\code{\link{combine}} NULL objects.}
\usage{combine.NULL(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





