\name{combine}
\alias{combine}
\title{\code{\link{combine}}}
\description{\code{\link{combine}} lists or character strings}
\usage{combine(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





