\name{CompareTrimmedNonEmptyLines}
\alias{CompareTrimmedNonEmptyLines}
\title{CompareTrimmedNonEmptyLines}
\description{helper function needed for comparison of doc files}
\usage{CompareTrimmedNonEmptyLines(s1, s2)}\arguments{
  \item{s1}{
}
  \item{s2}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





