\name{createObjects}
\alias{createObjects}
\title{createObjects}
\description{the function creates the environment object lists and expression by parsing all the code files
Is is factored out to make writing tests easier
since we often need the objects and the environment 
they inhabit }
\usage{createObjects(code)}\arguments{
  \item{code}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





