\name{decomment}
\alias{decomment}
\title{\code{\link{decomment}}}
\description{Remove comment \code{\link{prefix}} and join lines of code to form a
documentation string.}
\usage{decomment(comments)}
\arguments{
  \item{comments}{Character vector of \code{\link{prefix}}ed comment lines.}
}

\value{String without \code{\link{prefix}}es or newlines.}

\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





