\name{definition.from.source}
\alias{definition.from.source}
\title{definition from source}
\description{small helper to extract the definition of a doc entry from a bit of src code}
\usage{definition.from.source(doc, src, ...)}\arguments{
  \item{doc}{
}
  \item{src}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





