\name{do.not.generate}
\alias{do.not.generate}
\title{do not generate}
\description{Make a Parser Function used to indicate that certain Rd files
should not be generated.}
\usage{do.not.generate(...)}
\arguments{
  \item{\dots}{Character strings indicating Rd files without the .Rd suffix.}
}

\value{A Parser Function that will delete items from the outer
Documentation List.}

\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





