\name{empty.description}
\alias{empty.description}
\docType{data}
\title{empty description}
\description{Default DESCRIPTION, written if it doesn't exist.}
\usage{empty.description}
\format{
  The format is:
 chr [1, 1:7] "" "" "" "" "" "" ""
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:7] "Package" "Maintainer" "Author" "Version" ...
}
\details{
}
\source{
}
\references{
}
\examples{}