\name{escape_dots}
\alias{escape_dots}
\title{escape dots}
\description{Convert ... to \\dots}
\usage{escape_dots(arg)}
\arguments{
  \item{arg}{arg }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





