\name{exported}
\alias{exported}
\title{exported}
\description{a helper soon to read the NAMESPACE file, soon to be replaced by Rs own function}
\usage{exported(pattern, tD)}\arguments{
  \item{pattern}{
}
  \item{tD}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





