\name{exportedClasses}
\alias{exportedClasses}
\title{exportedClasses}
\description{get the exported Classes from the NAMESPACE file}
\usage{exportedClasses(tD)}\arguments{
  \item{tD}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





