\name{exportedDocumentableMeths}
\alias{exportedDocumentableMeths}
\title{exportedDocumentableMeths}
\description{#########################################################}
\usage{exportedDocumentableMeths(e, pkgDir)}\arguments{
  \item{e}{
}
  \item{pkgDir}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





