\name{exportedFunctions}
\alias{exportedFunctions}
\title{exportedFunctions}
\description{get the exported functions from the NAMESPACE file}
\usage{exportedFunctions(tD)}\arguments{
  \item{tD}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





