\name{exportedGenerics}
\alias{exportedGenerics}
\title{exportedGenerics}
\description{get the exported generic functions from the NAMESPACE file}
\usage{exportedGenerics(tD)}\arguments{
  \item{tD}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





