\name{extract.docs.fun}
\Rdversion{1.1}
\alias{extract.docs.fun}
\title{extract.docs.fun}
\description{Given a function name, return a list describing inline documentation in the source of that function.}
\usage{
extract.docs.fun(fun)
}
\arguments{
  \item{fun}{The function to examine.}
}
\details{
}
\value{Named list of character strings extracted from comments. Each name corresponds to an argument of the function or a special keyword (recognized by name.to.match.string) used to place the comments in the Rd file.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\note{
}


\seealso{
}