\name{extract.docs}
\alias{extract.docs}
\title{extract docs}
\description{produce doc link instances}
\usage{extract.docs(parsed, objs, on)}\arguments{
  \item{parsed}{
}
  \item{objs}{
}
  \item{on}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





