\name{fields}
\alias{fields}
\docType{data}
\title{fields}
\description{Necessary fields in DESCRIPTION, otherwise error.}
\usage{fields}
\format{
  The format is:
 chr [1:7] "Package" "Maintainer" "Author" "Version" "License" "Title" "Description"
}
\details{
}
\source{
}
\references{
}
\examples{}
