\name{findGeneric}
\alias{findGeneric}
\title{\code{\link{findGeneric}}}
\description{Copied from R-3.0.1, to support \code{\link{findGeneric}}.}
\usage{findGeneric(fname, envir)}
\arguments{
  \item{fname}{
}
  \item{envir}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





