\name{fixPackageFileNames}
\alias{fixPackageFileNames}
\title{\code{\link{fixPackageFileNames}}}
\description{Copied from R-3.0.1, to support \code{\link{fixPackageFileNames}}.}
\usage{fixPackageFileNames(list)}
\arguments{
  \item{list}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





