\name{forall.parsers}
\alias{forall.parsers}
\docType{data}
\title{forall parsers}
\description{Parsers for each object/function that are constructed
automatically. This is a named list, each element is a list of 2
elements: forfun/forall, then a parser function for an individual
object.}
\usage{forall.parsers}





