\name{GenHasAnyExposedMethod}
\alias{GenHasAnyExposedMethod}
\title{GenHasAnyExposedMethod}
\description{function used to check if a GenericFunction has any method where the whole signature consist of classes exported in the namespace}
\usage{GenHasAnyExposedMethod(genName, env, pkgDir)}\arguments{
  \item{genName}{
}
  \item{env}{
}
  \item{pkgDir}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





