\name{GenHasAnyMethodWithSrc}
\alias{GenHasAnyMethodWithSrc}
\title{GenHasAnyMethodWithSrc}
\description{function to check if we have a src reference for any of the methods of this generic
This helps to decide how the *-methods.Rd file should look like for this generic}
\usage{GenHasAnyMethodWithSrc(genName, env)}\arguments{
  \item{genName}{
}
  \item{env}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





