\name{GenHasSrc}
\alias{GenHasSrc}
\title{GenHasSrc}
\description{This function tells us if we can find a src reference for this generic}
\usage{GenHasSrc(genName, e)}\arguments{
  \item{genName}{
}
  \item{e}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





