\name{get_S3_primitive_generics}
\alias{get_S3_primitive_generics}
\title{get S3 primitive generics}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{get_S3_primitive_generics(include_group_generics = TRUE)}
\arguments{
  \item{include_group_generics}{
}
}



\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





