\name{getKnownS3generics}
\alias{getKnownS3generics}
\title{\code{\link{getKnownS3generics}}}
\description{Copied from R-3.0.1, to support \code{\link{getKnownS3generics}}.}
\usage{getKnownS3generics()}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





