\name{getMethodName}
\alias{getMethodName}
\title{getMethodName}
\description{#########################################################}
\usage{getMethodName(doc.link, e)}\arguments{
  \item{doc.link}{
}
  \item{e}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





