\name{getSource}
\alias{getSource}
\title{\code{\link{getSource}}}
\description{Extract a function's source code.}
\usage{getSource(fun.obj)}
\arguments{
  \item{fun.obj}{A function.}
}

\value{Source code lines as a character vector.}

\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





