\name{hiddenClasses}
\alias{hiddenClasses}
\title{hiddenClasses}
\description{#########################################################}
\usage{hiddenClasses(env, pkgDir)}\arguments{
  \item{env}{
}
  \item{pkgDir}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





