\name{inherit.docs}
\alias{inherit.docs}
\title{inherit docs}
\description{recursively add documentation inherited from doc.link parents }
\usage{inherit.docs(parsed, res, childName)}\arguments{
  \item{parsed}{a list of doc.link objects}
  \item{res}{the list of documentation to be extended}
  \item{childName}{the name of the object who possibly inherits}
}

\value{the possibly extended list of documentation}

\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





