\name{inlinedocs-package}
\alias{inlinedocs-package}
\alias{inlinedocs}
\docType{package}
\title{Convert inline comments to documentation}
\description{Generates Rd files from R source code with comments.
The main features of the default syntax are that
(1) docs are defined in comments near the relevant code,
(2) function argument names are not repeated in comments, and
(3) examples are defined in R code, not comments.
It is also easy to define a new syntax.}
\details{
\tabular{ll}{Package: \tab inlinedocs\cr
Title: \tab Convert inline comments to documentation\cr
Type: \tab Package\cr
Version: \tab 2014.6.3\cr
Authors@R: \tab c(person("Toby Dylan","Hocking",role=c("aut","cre"),
email="toby@sg.cs.titech.ac.jp"),
person("Keith","Ponting",role="aut"),
person("Thomas","Wutzler",role="aut"),
person("Philippe","Grosjean",role="aut"),
person("Markus","M端ller",role="aut"),
person("R Core Team",role=c("ctb","cph")))\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
Depends: \tab methods, utils, R (>= 2.9)\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
Encoding: \tab UTF-8\cr
Author: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus M端ller [aut], R Core Team [ctb, cph]\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus M端ller [aut], R Core Team [ctb, cph]\cr}
}
\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus M端ller [aut], R Core Team [ctb, cph]}

\keyword{ package }


