\name{kill.prefix.whitespace}
\alias{kill.prefix.whitespace}
\title{kill prefix whitespace}
\description{Figure out what the whitespace preceding the example code is, and
then delete that from every line.}
\usage{kill.prefix.whitespace(ex)}
\arguments{
  \item{ex}{character vector of example code lines.}
}

\value{Character vector of code lines with preceding whitespace removed.}

\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





