\name{methodDocName}
\alias{methodDocName}
\title{methodDocName}
\description{creates the actual *.Rd filename for a method from its signature and the generic it implements}
\usage{methodDocName(genName, sig)}\arguments{
  \item{genName}{
}
  \item{sig}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





