\name{MethodSignatureHasOnlyExportedClasses}
\alias{MethodSignatureHasOnlyExportedClasses}
\title{MethodSignatureHasOnlyExportedClasses}
\description{check if all the classes in the signature are exported in the NAMESPACE file.
This information is needed to decide which Methods we want to document in cases
where the documentations is restricted to the exported NAMESPACE}
\usage{MethodSignatureHasOnlyExportedClasses(MethodDefinition, env, 
    pkgDir)}\arguments{
  \item{MethodDefinition}{
}
  \item{env}{
}
  \item{pkgDir}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





