\name{MethodsWithSrcRefForGen}
\alias{MethodsWithSrcRefForGen}
\title{MethodsWithSrcRefForGen}
\description{Not all methods for a Generic are defined in the src we want to document.
This function helps to find the methods we want.}
\usage{MethodsWithSrcRefForGen(genName, env)}\arguments{
  \item{genName}{
}
  \item{env}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





