\name{methodTable}
\alias{methodTable}
\title{methodTable}
\description{#########################################################}
\usage{methodTable(exprs, e)}\arguments{
  \item{exprs}{
}
  \item{e}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





