\name{methSig}
\alias{methSig}
\title{methSig}
\description{Extract the definition as text from an MethodDefinition object}
\usage{methSig(MethodDefinition)}\arguments{
  \item{MethodDefinition}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





