\name{methSrc}
\alias{methSrc}
\title{methSrc}
\description{get at the src of a method given as  an MethodDefinition object}
\usage{methSrc(MethodDefinition)}\arguments{
  \item{MethodDefinition}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





