\name{mm.examples.from.testfile}
\alias{mm.examples.from.testfile}
\title{mm examples from testfile}
\description{extract examples from external files }
\usage{mm.examples.from.testfile(name, inlinedocs.exampleDir, inlinedocs.exampleTrunk, 
    ...)}\arguments{
  \item{name}{
}
  \item{inlinedocs.exampleDir}{
}
  \item{inlinedocs.exampleTrunk}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





