\name{mmPromptMethods}
\alias{mmPromptMethods}
\title{mmPromptMethods}
\description{#########################################################}
\usage{mmPromptMethods(genName, filename = NULL, exportedMeths, where)}\arguments{
  \item{genName}{
}
  \item{filename}{
}
  \item{exportedMeths}{
}
  \item{where}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





