\name{modify.Rd.file}
\alias{modify.Rd.file}
\title{modify Rd file}
\description{Add inline documentation from comments to an Rd file
automatically-generated by package.skeleton.}
\usage{modify.Rd.file(N, pkg, docs)}
\arguments{
  \item{N}{Name of function/file to which we will add documentation.}
  \item{pkg}{Package name.}
  \item{docs}{Named list of documentation in extracted comments.}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





