\name{pe}
\alias{pe}
\title{pe}
\description{This function is used to print out a value of an quoted expression together with it and is helpful for debugging}
\usage{pe(string, env)}\arguments{
  \item{string}{
}
  \item{env}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





