\name{pp}
\alias{pp}
\title{print out an }
\description{This function is used to print out a value of a variable together with its name and is helpful for debugging}
\usage{pp(string, env)}\arguments{
  \item{string}{
}
  \item{env}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





