\name{print.allfun}
\alias{print.allfun}
\title{\code{\link{print allfun}}}
\description{Print method for functions constructed using \code{\link{forall}}.}
\usage{\method{print}{allfun}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <tdhock5@gmail.com> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





