\name{remove.attributes.file}
\alias{remove.attributes.file}
\title{remove attributes file}
\description{Remove attributes from all R objects defined in a file. This may
be useful if for some reason we want to remove the examples from
the ex attribute of functions.}
\usage{remove.attributes.file(f)}
\arguments{
  \item{f}{File to parse and then write objects without attributes.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





