\name{save.test.result}
\alias{save.test.result}
\title{save test result}
\description{For unit tests, this is an easy way of getting a text
representation of the list result of \code{\link{extract.docs.file}}.}
\usage{save.test.result(f)}
\arguments{
  \item{f}{R code file with inlinedocs to process with \code{\link{extract.docs.file}}.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





