\name{sigString}
\alias{sigString}
\title{sigString}
\description{#########################################################}
\usage{sigString(sig)}\arguments{
  \item{sig}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





