\name{title.from.firstline}
\alias{title.from.firstline}
\title{title from firstline}
\description{extract the title from the first line of a function definition}
\usage{title.from.firstline(src, ...)}\arguments{
  \item{src}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





