\name{trimmedNonEmptyLines}
\alias{trimmedNonEmptyLines}
\title{trimmedNonEmptyLines}
\description{a helper function to make text comparison of actual and expacted output less painfull by removing whitespace}
\usage{trimmedNonEmptyLines(s)}\arguments{
  \item{s}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





