\name{whole.word}
\alias{whole.word}
\title{whole word}
\description{Regex for a whole word to code/link tags.}
\usage{whole.word(...)}
\arguments{
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





