\name{writeClassRdFiles}
\alias{writeClassRdFiles}
\title{writeClassRdFiles}
\description{##############################################################}
\usage{writeClassRdFiles(environment, pkgDir, path, inlinedocs.documentNamespaceOnly)}\arguments{
  \item{environment}{
}
  \item{pkgDir}{
}
  \item{path}{
}
  \item{inlinedocs.documentNamespaceOnly}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





