\name{writeFunctionRdFiles}
\alias{writeFunctionRdFiles}
\title{writeFunctionRdFiles}
\description{##############################################################
At the moment only empty files are written (and later filled my modify.Rd.file) although the complete information is
##############################################################
already present at this stage. Future versions will finish the job right here, rendering modify.Rd file unnecassary
##############################################################
but only after all the other stuff can be written directly too.}
\usage{writeFunctionRdFiles(e, pkgDir, path, inlinedocs.documentNamespaceOnly)}\arguments{
  \item{e}{
}
  \item{pkgDir}{
}
  \item{path}{
}
  \item{inlinedocs.documentNamespaceOnly}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





