\name{writeMethodRdFiles}
\alias{writeMethodRdFiles}
\title{writeMethodRdFiles}
\description{##############################################################}
\usage{writeMethodRdFiles(e, pkgDir, path, inlinedocs.documentNamespaceOnly = FALSE)}\arguments{
  \item{e}{
}
  \item{pkgDir}{
}
  \item{path}{
}
  \item{inlinedocs.documentNamespaceOnly}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





