\name{writePackageRdFile}
\alias{writePackageRdFile}
\title{writePackageRdFile}
\description{##############################################################}
\usage{writePackageRdFile(pkgDir, name, path)}\arguments{
  \item{pkgDir}{
}
  \item{name}{
}
  \item{path}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





