\name{[_method__listOfMethods_logical}
\alias{[_method__listOfMethods_logical}
\title{[ method  listOfMethods logical}
\description{overload the [] operator for objects of class "listOfMethods"}
\arguments{
  \item{x}{
}
  \item{i}{
}
  \item{j}{
}
  \item{\dots}{
}
  \item{drop}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





