\name{combine.character}
\alias{combine.character}
\title{combine character}
\description{combine character strings by pasting them together}
\usage{
combine.character(x, y)
}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>, Keith Ponting <k.ponting@aurix.com>}





