\name{combine.character}
\alias{combine.character}
\title{combine character}
\description{combine character strings by pasting them together}
\usage{
\method{combine}{character}(x, y)
}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





