\name{combine.character}
\alias{combine.character}
\title{combine character}
\description{\code{\link{combine}} character strings by pasting them together}
\usage{\method{combine}{character}(x, 
    y)}
\arguments{
  \item{x}{x }
  \item{y}{y }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





