\name{combine.list}
\alias{combine.list}
\title{combine list}
\description{combine lists by adding elements or adding to existing elements}
\usage{
combine.list(x, y)
}
\arguments{
  \item{x}{
}
  \item{y}{
}
}

\value{A list, same type as x, but with added elements from y.}

\author{Toby Dylan Hocking <toby.hocking@inria.fr>, Keith Ponting <k.ponting@aurix.com>}





