\name{combine.list}
\alias{combine.list}
\title{combine list}
\description{combine lists by adding elements or adding to existing elements}
\usage{\method{combine}{list}(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}

\value{A list, same type as x, but with added elements from y.}

\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





