\name{combine.list}
\alias{combine.list}
\title{combine list}
\description{\code{\link{combine}} lists by adding elements or adding to existing elements}
\usage{\method{combine}{list}(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}

\value{A list, same type as \code{x}, but with added elements from \code{y}.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





