\name{combine.list}
\alias{combine.list}
\title{combine list}
\description{combine lists by adding elements or adding to existing elements}
\usage{\method{combine}{list}(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}

\value{A list, same type as x, but with added elements from y.}

\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





