\name{combine.NULL}
\alias{combine.NULL}
\title{combine NULL}
\description{\code{\link{combine}} NULL objects.}
\usage{\method{combine}{NULL}(x, y)}
\arguments{
  \item{x}{
}
  \item{y}{
}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





