\name{decomment}
\alias{decomment}
\title{decomment}
\description{Remove comment prefix and join lines of code to form a
documentation string.}
\usage{decomment(comments)}
\arguments{
  \item{comments}{Character vector of prefixed comment lines.}
}

\value{String without prefixes or newlines.}

\author{Inlinedocs development team with contributions from R Core Team}





