\name{decomment}
\alias{decomment}
\title{decomment}
\description{Remove comment prefix and join lines of code to form a
documentation string.}
\usage{
decomment(comments)
}
\arguments{
  \item{comments}{Character vector of prefixed comment lines.}
}
\details{
}
\value{String without prefixes or newlines.}
\references{
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>, Keith Ponting <k.ponting@aurix.com>}
\note{
}


\seealso{
}
\examples{}