\name{decomment}
\alias{decomment}
\title{decomment}
\description{Remove comment \code{\link{prefix}} and join lines of code to form a
documentation string.}
\usage{decomment(comments)}
\arguments{
  \item{comments}{Character vector of prefixed comment lines.}
}

\value{String without prefixes or newlines.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





