\name{findGeneric}
\alias{findGeneric}
\title{findGeneric}
\description{Copied from R-3.0.1, to support findGeneric.}
\usage{findGeneric(fname, envir)}
\arguments{
  \item{fname}{
}
  \item{envir}{
}
}



\author{Inlinedocs development team with contributions from R Core Team}





