\name{findGeneric}
\alias{findGeneric}
\title{findGeneric}
\description{Copied from R-3.0.1, to support findGeneric.}
\usage{findGeneric(fname, envir)}
\arguments{
  \item{fname}{
}
  \item{envir}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





