\name{fixPackageFileNames}
\alias{fixPackageFileNames}
\title{fixPackageFileNames}
\description{Copied from R-3.0.1, to support fixPackageFileNames.}
\usage{fixPackageFileNames(list)}
\arguments{
  \item{list}{
}
}



\author{Inlinedocs development team with contributions from R Core Team}





