\name{fixPackageFileNames}
\alias{fixPackageFileNames}
\title{fixPackageFileNames}
\description{Copied from R-3.0.1, to support fixPackageFileNames.}
\usage{fixPackageFileNames(list)}
\arguments{
  \item{list}{
}
}



\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





