\name{fixPackageFileNames}
\alias{fixPackageFileNames}
\title{fixPackageFileNames}
\description{Copied from R-3.0.1, to support fixPackageFileNames.}
\usage{fixPackageFileNames(list)}
\arguments{
  \item{list}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





