\name{forfun}
\alias{forfun}
\title{forfun}
\description{For each function in the package, do something.
For each object in the package that satisfies the criterion
For each function in the package, do something.
checked by subfun, parse source using FUN and return the resulting
For each function in the package, do something.
documentation list.}
\usage{forfun(FUN)}\arguments{
  \item{FUN}{Function to apply to each function in the package.}
}

\value{A Parser Function.}

\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





