\name{forfun}
\alias{forfun}
\title{forfun}
\description{For each function in the package, do something.}
\usage{forfun(FUN)}
\arguments{
  \item{FUN}{FUN }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





