\name{get_internal_S3_generics}
\alias{get_internal_S3_generics}
\title{get internal S3 generics}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{get_internal_S3_generics(primitive = TRUE)}
\arguments{
  \item{primitive}{
}
}



\author{Inlinedocs development team with contributions from R Core Team}





