\name{get_internal_S3_generics}
\alias{get_internal_S3_generics}
\title{get internal S3 generics}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{get_internal_S3_generics(primitive = TRUE)}
\arguments{
  \item{primitive}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





