\name{getKnownS3generics}
\alias{getKnownS3generics}
\title{getKnownS3generics}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{getKnownS3generics()}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





