\name{getSource}
\alias{getSource}
\title{getSource}
\description{Extract a function's source code.}
\usage{getSource(fun.obj)}
\arguments{
  \item{fun.obj}{A function.}
}

\value{Source code lines as a character vector.}

\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





