\name{getSource}
\alias{getSource}
\title{getSource}
\description{Extract a function's source code.}
\usage{getSource(fun.obj)}
\arguments{
  \item{fun.obj}{A function.}
}

\value{Source code lines as a character vector.}

\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





