\name{inlinedocs-package}
\Rdversion{1.1}
\alias{inlinedocs-package}
\docType{package}
\title{Convert inline comments to documentation}
\description{An alternative to Roxygen that focuses on the Don't Repeat Yourself principle and keeping code and documentation close together.}
\details{
\tabular{ll}{Package: \tab inlinedocs\cr
Title: \tab Convert inline comments to documentation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-29\cr
Author: \tab Toby Dylan Hocking\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@etu.upmc.fr>\cr
Description: \tab An alternative to Roxygen that focuses on the Don't Repeat Yourself principle and keeping code and documentation close together.\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}