\name{inlinedocs-package}
\alias{inlinedocs-package}
\title{Convert inline comments to documentation}
\description{Generates Rd files from R source code with comments, providing for quick, sustainable package development. The syntax keeps code and documentation close together, and is inspired by the Don't Repeat Yourself principle.}

\author{Inlinedocs development team <inlinedocs-support@lists.r-forge.r-project.org>}

\keyword{ package }


