\name{inlinedocs-package}
\Rdversion{1.1}
\alias{inlinedocs-package}
\docType{package}
\title{inlinedocs-package}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab inlinedocs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-31\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>}
\references{
}
\keyword{ package }
\seealso{
}