\name{inlinedocs-package}
\Rdversion{1.1}
\alias{inlinedocs-package}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab inlinedocs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-08-01\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
}
\keyword{ package }
\seealso{
}