\name{inlinedocs-package}
\alias{inlinedocs-package}
\docType{package}
\title{Convert inline comments to documentation}
\description{Generates Rd files from R source code with comments, providing for quick, sustainable package development. The syntax keeps code and documentation close together, and is inspired by the Don't Repeat Yourself principle.}
\details{
\tabular{ll}{Package: \tab inlinedocs\cr
Title: \tab Convert inline comments to documentation\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2010-10-20\cr
Author: \tab Inlinedocs development team\cr
Maintainer: \tab Inlinedocs development team <inlinedocs-support@lists.r-forge.r-project.org>\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
Depends: \tab methods, utils\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Inlinedocs development team <inlinedocs-support@lists.r-forge.r-project.org>}

\keyword{ package }


