\name{inlinedocs-package}
\alias{inlinedocs-package}
\alias{inlinedocs}
\docType{package}
\title{Convert inline comments to documentation}
\description{Generates Rd files from R source code with comments.
The main features of the default syntax are that
(1) docs are defined in comments near the relevant code,
(2) function argument names are not repeated in comments, and
(3) examples are defined in R code, not comments.
It is also easy to define a new syntax.}
\details{
\tabular{ll}{Package: \tab inlinedocs\cr
Title: \tab Convert inline comments to documentation\cr
Type: \tab Package\cr
Version: \tab 2013.8.22\cr
Author: \tab Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
Depends: \tab methods, utils, R (>= 2.9)\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}

\keyword{ package }


