\name{inlinedocs-package}
\Rdversion{1.1}
\alias{inlinedocs-package}
\docType{package}
\title{Convert inline comments to documentation}
\description{Generates Rd files from R source code with comments, providing for quick, sustainable package development. The syntax keeps code and documentation close together, and is inspired by the Don't Repeat Yourself principle.}
\details{
\tabular{ll}{Package: \tab inlinedocs\cr
Title: \tab Convert inline comments to documentation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-29\cr
Author: \tab Toby Dylan Hocking, Keith Ponting\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@etu.upmc.fr>, Keith Ponting <k.ponting@aurix.com>\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Toby Dylan Hocking <toby.hocking@etu.upmc.fr>, Keith Ponting <k.ponting@aurix.com>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}
