\name{is_primitive_in_base}
\alias{is_primitive_in_base}
\title{is primitive in base}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{is_primitive_in_base(fname)}
\arguments{
  \item{fname}{
}
}



\author{Inlinedocs development team with contributions from R Core Team}





