\name{is_primitive_in_base}
\alias{is_primitive_in_base}
\title{is primitive in base}
\description{Copied from R-3.0.1, to support getKnownS3generics.}
\usage{is_primitive_in_base(fname)}
\arguments{
  \item{fname}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





