\name{kill.prefix.whitespace}
\alias{kill.prefix.whitespace}
\title{kill prefix whitespace}
\description{Figure out what the whitespace preceding the example code is, and
then delete that from every line.}
\usage{kill.prefix.whitespace(ex)}
\arguments{
  \item{ex}{character vector of example code lines.}
}

\value{Character vector of code lines with preceding whitespace removed.}

\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





