\name{make.package.and.check}
\alias{make.package.and.check}
\title{make package and check}
\description{Make a package by processing f with a standard DESCRIPTION, and
stop if there are errors or warnings.}
\usage{
make.package.and.check(f, parsers = default.parsers)
}
\arguments{
  \item{f}{
}
  \item{parsers}{
}
}



\author{Inlinedocs development team <inlinedocs-support@lists.r-forge.r-project.org>}





