\name{print.allfun}
\alias{print.allfun}
\title{print allfun}
\description{Print method for functions constructed using forall.}
\usage{\method{print}{allfun}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Inlinedocs development team with contributions from R Core Team}





