\name{print.allfun}
\alias{print.allfun}
\title{print allfun}
\description{Print method for functions constructed using forall.}
\usage{
\method{print}{allfun}(x, ...)
}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





