\name{print.allfun}
\alias{print.allfun}
\title{print allfun}
\description{Print method for functions constructed using \code{\link{forall}}.}
\usage{\method{print}{allfun}(x, ...)}
\arguments{
  \item{x}{x }
  \item{\dots}{\dots }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





