\name{print.allfun}
\alias{print.allfun}
\title{print allfun}
\description{Print method for functions constructed using forall.}
\usage{\method{print}{allfun}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





