\name{removeAliasesfrom.Rd.file}
\alias{removeAliasesfrom.Rd.file}
\title{removeAliasesfrom Rd file}
\description{remove aliases to methodnames from the Rd file of a class
automatically-generated by package.skeleton.}
\usage{removeAliasesfrom.Rd.file(N, pkg, code)}
\arguments{
  \item{N}{Name of function/file to which we will add documentation.}
  \item{pkg}{Package name.}
  \item{code}{The code of the package }
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





