\name{replace.one}
\alias{replace.one}
\title{replace one}
\description{Do find and replace for one element of an inner documentation list
on 1 Rd file.}
\usage{replace.one(torep, REP, txt)}
\arguments{
  \item{torep}{tag to find.}
  \item{REP}{contents of tag to put inside.}
  \item{txt}{text in which to search.}
}



\author{Inlinedocs development team
<inlinedocs-support@lists.r-forge.r-project.org>}





