\name{replace.one}
\alias{replace.one}
\title{replace one}
\description{Do find and replace for one element of an inner documentation list
on 1 Rd file.}
\usage{replace.one(torep, REP, txt)}
\arguments{
  \item{torep}{tag to find.}
  \item{REP}{contents of tag to put inside.}
  \item{txt}{text in which to search.}
}



\author{Toby Dylan Hocking <toby@sg.cs.titech.ac.jp> [aut, cre], Keith Ponting [aut], Thomas Wutzler [aut], Philippe Grosjean [aut], Markus Müller [aut], R Core Team [ctb, cph]}





