\name{save.test.result}
\alias{save.test.result}
\title{save test result}
\description{For unit tests, this is an easy way of getting a text
representation of the list result of extract.docs.file.}
\usage{
\method{save}{test.result}(f)
}
\arguments{
  \item{f}{R code file with inlinedocs to process with extract.docs.file.}
}



\author{Inlinedocs development team <inlinedocs-support@lists.r-forge.r-project.org>}





